third_party/WebKit/LayoutTests/http/tests/devtools/editor/text-editor-mark-clean-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
This test checks TextEditorModel.markClean/isClean methods


Running: testMarkiningInitialStateAsClean
Initial state: clean=false
After marking clean: clean=true
EDIT; clean=false
UNDO; clean=true
REDO; clean=false
UNDO; clean=true
EDIT; clean=false
UNDO; clean=true

Running: testMiddleStateAsClean
Initial state: clean=true
EDIT; clean=false
EDIT; clean=false
EDIT; clean=false
After marking clean: clean=true
EDIT; clean=false
UNDO; clean=true
UNDO; clean=false
UNDO; clean=false
UNDO; clean=false
REDO; clean=false
REDO; clean=false
REDO; clean=true
REDO; clean=false
UNDO; clean=true
UNDO; clean=false
EDIT; clean=false
UNDO; clean=false
UNDO; clean=false
